% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{filters.trade-package}
\alias{filters.trade}
\alias{filters.trade-package}
\title{filters.trade Package}
\description{
A filter system for selecting trading instruments. \code{add_filter()} and
\code{apply_filter()} are the two entry points of the package.
}
\author{
\strong{Maintainer}: Joe Zhu \email{sha.joe.zhu@gmail.com}

Authors:
\itemize{
  \item Joe Zhu \email{sha.joe.zhu@gmail.com}
}

}
