% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{get_filter}
\alias{get_filter}
\title{Get a Filter Definition}
\usage{
get_filter(id)
}
\arguments{
\item{id}{\code{character}. The filter ID}
}
\value{
A \code{list} with elements \code{title}, \code{target} and \code{condition}
}
\description{
Get a Filter Definition
}
\examples{
add_filter(
  id = "LARGECAP",
  title = "Large Cap",
  target = "TICKERS",
  condition = MARKET_CAP >= 1e10,
  overwrite = TRUE
)
get_filter("LARGECAP")

## Filter `FOO` does not exist
try(get_filter("FOO"))

}
